Procedurally generate `raw.rs`
This commit adds a script that uses the `witx` crate to parse the
official WASI specification definitions and generate Rust bindings for
the upstream APIs. The intention is that this isn't actually a breaking
change in general (but it is in a few places that switch from `*mut
c_void` to `*mut u8`).

A CI check is also added to ensure that the in-repository `raw.rs` is
the same as what should be generated, which allows shipping this to
crates.io without actually parsing `*.witx` at build time.