Procedurally generate the entire crate
This commit switches to using the `*.witx` file to procedurally generate
this entire crate. Both the raw low-level function bindings as well as
the higher level wrappers which use types like slices and `&str` are
procedurally generated.

Along the way a number of various other modifications and improvements
were made:

* Documentation is carried over from the `*.witx` file to bindings generated.
* The WASI bindings are updated to `wasi_snapshot_preview1`.
* The crate version is bumped, and now contains build metadata showing
  what version of the WASI spec is implemented.
* A dedicated opaque `Error` type has been added which implements traits
  like `Debug` and `Display` which are WASI-specific. It also
  additionally implements the `Error` trait if enabled.
* The crate's module structure has been removed, with bindings now
  appearing at the top level instead of in a `wasi_snapshot_preview1`
  submodule.
* The `__wasi` and `__WASI` prefixes on names have all been removed.
* The names of enums values and flag values are now `{ENUM}_{FLAG}` to
  disambiguate same-named flags between enum/flag types.

The entire crate is now procedurally generated but this means that a few
prior convenience wrappers no longer exist and/or are much more raw then
they were before:

* `args_get` - this simply takes raw buffers instead of a closure
* `environ_get` - takes raw buffers like above
* `fd_prestat_dir_name` - takes a raw pointer to write to instead of a
  mutable buffer
* `fd_readdir` - same as above
* `path_readlink` - same as above
* `random_get` - same as above
* `poll_oneoff` - this seems like a bit of a weird API, but no longer
  takes two arrays and only works in raw pointers

I think this means we'll probably want future wasi spec work to codify a
"mutable output buffer" as first class concept for most of the APIs
above, and I think that would make most of them nicer to use.

Closes #20