Separate generated code into a private submodule lib_generated
Move all the generated code into `lib_generated.rs`, load that as a
private submodule, and `pub use lib_generated::*;`. This allows moving
the crate documentation, `#![no_std]`, and `mod error;` out of the
generated code and into the now-non-generated `lib.rs`. This makes the
generated code capable of functioning with `include!` in the future.

Modify the generated `strerror` to use `pub(crate)` so that the `error`
module can reference it.

These changes have no effect on the visible API or documentation of the
crate.