Add a wasi-ephemeral crate to build with ephemeral WASI interfaces
Developers working on WASI interfaces can use the wasi-ephemeral crate
as a path dependency or git dependency, to build and run with the
ephemeral WASI interface rather than the latest snapshot.
(wasi-ephemeral uses publish=false and will not be uploaded to
crates.io.)

wasi-ephemeral builds with a build.rs script, so that it automatically
rebuilds on changes to the witx file.

wasi-ephemeral also supports building with an arbitrary witx file, using
the WASI_EPHEMERAL_WITX environment variable.