Update wit-bindgen and bump to 0.13.0
This commit updates wit-bindgen to 0.21.0 which brings in some minor
updates to generated code in addition to a new `wit-bindgen-rt` crate
dependency to use as well. This avoids the need to depend on the
`wit-bindgen` crate at all through some updated options to the CLI
bindings generator.

Additionally this bumps the crate to 0.13.0 to release these updated
change in addition to the refactoring for the export-related macros
previously.