Don't generate bindings with a synthetic package (#115)
This commit addresses an issue in Rust nightly with the Rust standard
library. Notably what's happening is that the `wasi` crate is using a
synthetic world/package called `rust:wasi/bindings` to generate
bindings. This world includes `wasi:cli/imports` and
`wasi:http/imports`. What's happening is that the Rust standard library
is using one version of the `wasi` crate with one version of WASI
bindings, but user crates might use something different. This fails to
link because `rust:wasi/bindings` is asserted to be the same no matter
what and there's two different `wasi` crate versions feeding into this
which cause conflicts.

The fix in this commit is to skip using `rust:wasi` as a package and
generate bindings directly from WASI packages themselves. These bindings
correctly have versions in them and are versioned appropriately so
there's no clash and merging proceeds as expected.

This required moving the bindings for `wasi:http/imports` into the
generation of bindings for `wasi:http/proxy`, but that doesn't actually
change the structure of the crate just how the generated bindings work
out.