Develop bindings for WASI versions in parallel (#118)
* Develop bindings for WASI versions in parallel

This commit is a reorganization of this crate to develop the
WASIp{1,2,3} crate bindings in parallel with each other in separate
crates. Historically the WASIp1 support ceased with the publication of
0.12.0 of the `wasi` crate which was the first release generated by
`wit-bindgen` suitable for use with WASIp2. The rough idea was to do the
same with WASIp3 at some point, but today it was realized that this
probably won't work.

Specifically the version number of the `wasi` crate right now is tied to
not only the WASI API version that's being bound but addition to the
`wit-bindgen` that is generating bindings. If either one updates in a
breaking way then the crate number here needs to be bumped. This means
that if we want to develop both WASIp2 and WASIp3 at the same time we'd
have to somehow reserve a space of future version numbers for the WASIp2
track which would be a bit awkward. Additionally it's a bit awkward to
have non-`main` development of WASIp2, a supported target, when WASIp3
wants to be added.

This new crate organization is intended to solve this concerns and looks
like so:

* Each version of WASI now has its own crate. For example there's now a
  `wasip1` crate and a `wasip2` crate. Eventually there will be a
  `wasip3` crate.
* A version-specific WASI crate is developed on `main` of this
  repository and gets CI run, auto-generated bindings verified, etc.
  Versions will be published as necessary for each version of each crate.
* The `wasi` crate is now a small wrapper around the latest stable WASI
  release available. Currently this is a stable reexport of the `wasip2`
  crate.

Various bits of README documentation have all been updated to reflect
this new structure. The intention is that in the near-ish future a
`wasip3` crate will be added which will contain auto-generated bindings
for WASIp3 and its upcoming snapshots. This new `wasip3` crate would not
be integrated into the `wasi` crate and the `wasi` crate would continue
to reexport `wasip2`. In this manner we can develop `wasip3` more easily
alongside the `wasip2` crate.

* Fixes for CI

* Update msrv to 78

* Cancel in-flight CI

* Update msrv

* Download precompiled binary of wit-bindgen

* Downgrade one point

* Remove version suffix