drop write end of stream before writing trailers
This tells the receiver that there is no body content remaining and that it
should switch to reading the trailers, if any.  Otherwise, the sender and
receiver will deadlock, with the sender waiting to rendezvous its write to the
trailers future and the receiver waiting to receive more content bytes.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>